--- src/pages/home/subPages/dashboard/components/CurrentTagPath/index.ts
@@ -1,1 +1,0 @@
-export { default as CurrentTagPath } from './CurrentTagPath';
