+++ src/pages/home/subPages/dashboard/components/DashboardHeader/index.ts
@@ -1,0 +1,1 @@
+export { default as DashboardHeader } from './DashboardHeader';
