--- src/pages/home/subPages/dashboard/components/MemoSectionList/MemoSection/MemoSectionHeader/TagWithOptions/index.ts
@@ -1,1 +1,0 @@
-export { default as TagWithOptions } from './TagWithOptions';
