+++ src/pages/home/subPages/dashboard/components/TagOptions/index.ts
@@ -1,0 +1,1 @@
+export { default as TagOptions } from './TagOptions';
