+++ src/pages/home/subPages/dashboard/components/index.ts
@@ -1,4 +1,5 @@
-export * from './CurrentTagPath';
+export * from './DashboardHeader';
 export * from './MemoSectionList';
 export * from './MemoEditModal';
 export * from './TagEditModal';
+export * from './TagOptions';
